requirements.txt
torch
fairscale
fire
sentencepiece
pyyaml
transformers
torchtext
tokenizers
datasets
matplotlib

modelsummary
evaluate
tqdm
numpy